Implementação do serviço de atualização de status de procurações vencidas e agendamento do serviço no taskManager.

+++ server/services/ProcuracaoService.js
@@ -1,4 +1,5 @@
 //@ts-check
+const { ProcuracaoDaoImpl } = require("../infrastructure/ProcuracaoDaoImpl");
 const ProcuradorRepository = require("../repositories/ProcuradorRepository");
 const { Repository } = require("../repositories/Repository");
 const { ProcuradorService } = require("./ProcuradorService");
@@ -24,21 +25,24 @@ class ProcuracaoService {
     }
 
     /**
+     * Apaga a procuração, com a opção de softDelete, que mantém a procuração no sistema, mas revoga as permissões de usuário/procuradores
+     * O soft delete é utilizado pelo TaskManager para atualizar automaticamente procurações que estiverem vencidas.
+     *      *
      * @param {number} id
+     * @param {boolean} [softDelete]
      * @returns {Promise<object|false>} object {updatedProcuradores, codigoEmpresa} | false
      */
-    static async deleteProcuracao(id) {
+    static async deleteProcuracao(id, softDelete = false) {
         const procuracaoRepository = new Repository('procuracoes', 'procuracao_id')
-
         try {
             const procuracoes = await procuracaoRepository.find(id)
             if (!procuracoes.length) {
                 return false
             }
 
-            const result = await procuracaoRepository.delete(id)
-            if (!result) {
-                return false
+            if (!softDelete) {
+                const result = await procuracaoRepository.delete(id)
+                if (!result) return false
             }
 
             const { codigo_empresa: codigoEmpresa, procuradores: procuradorIds } = procuracoes[0]
@@ -57,6 +61,32 @@ class ProcuracaoService {
         }
     }
 
+    /**
+     * O soft delete é utilizado pelo TaskManager para atualizar automaticamente procurações que estiverem vencidas.
+     * @returns {Promise<number|false|undefined>} Undefined em caso de erro
+     */
+    static async softDeleteExpired() {
+        try {
+            const expiredProcuracoes = await ProcuracaoDaoImpl.getExpiredProcuracoes()
+            const expiredCount = expiredProcuracoes.length
+            if (expiredCount === 0) {
+                return false
+            }
+
+            const procuracaoRepository = new Repository('procuracoes', 'procuracao_id')
+            const updates = expiredProcuracoes.map(p => ({ procuracao_id: p.procuracao_id, status: 'vencida' }))
+            await procuracaoRepository.updateMany(updates)
+
+            for (const { procuracao_id } of expiredProcuracoes) {
+                const softDelete = true
+                ProcuracaoService.deleteProcuracao(procuracao_id, softDelete)
+            }
+
+            return expiredCount
+        } catch (error) {
+            console.log("🚀 ~ file: ProcuracaoService.js:87 ~ ProcuracaoService ~ softDeleteExpired ~ error:", error)
+        }
+    }
 }
 
 module.exports = { ProcuracaoService }
